requirements.txt
streamlit>=1.30.0
crewai>=0.108.0
crewai-tools>=0.38.1
google-generativeai>=0.8.0
langchain-google-genai>=2.1.0
langchain>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
jinja2>=3.1.0
fpdf2>=2.7.0
